Pick good changes from beta combat implementation.

+++ db/src/models/monster/mod.rs
@@ -20,10 +20,10 @@ pub struct Monster {
     pub id: i32,
     pub name: String,
     pub monster_type: MonsterType,
-    pub attack: i32,
-    pub hp: i32,
-    pub speed: i32,
-    pub defense: i32,
+    pub attack: f64,
+    pub hp: f64,
+    pub speed: f64,
+    pub defense: f64,
 }
 
 impl Model for Monster {
